backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["gavin-k-cho <gavin.k.cho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
flask-pymongo = "^2.3.0"
apscheduler = "^3.10.4"
requests = "^2.31.0"
flask-cors = "^4.0.0"
flask-jwt-extended = "^4.6.0"
neo4j = "^5.20.0"
flask-session = "^0.8.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
